requirements.txt
anthropic>=0.45.0
playwright>=1.49.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
python-dateutil>=2.9.0.post0
pydantic>=2.9.2
pandas>=2.2.3
python-dotenv>=1.0.1
httpx>=0.27.2
judgeval

